tcc/requirements.txt
optuna
zarr
pyyaml
opencv-python
pandas
numpy
tqdm
plotly
imagecodecs
numcodecs
hydra-core
kaleido
